+++ bin/app.php
@@ -1,6 +1,11 @@
 #!/usr/bin/env php
 <?php
 
+use Doctrine\Migrations\Configuration\Configuration;
+use Doctrine\Migrations\Tools\Console\Helper\ConfigurationHelper;
+use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\Tools\Console\ConsoleRunner;
+use Doctrine\ORM\Tools\Console\Helper\EntityManagerHelper;
 use Symfony\Component\Console;
 
 chdir(dirname(__DIR__));
@@ -11,9 +16,23 @@ $container = require 'config/container.php';
 
 $cli = new Console\Application('Application console');
 
+/** @var EntityManagerInterface $entityManager */
+$entityManager = $container->get(EntityManagerInterface::class);
+$connection = $entityManager->getConnection();
+
+$configuration = new Configuration($connection);
+$configuration->setMigrationsDirectory('src/App/Migration');
+$configuration->setMigrationsNamespace('Migration');
+
+$cli->getHelperSet()->set(new EntityManagerHelper($entityManager), 'em');
+$cli->getHelperSet()->set(new ConfigurationHelper($connection, $configuration), 'configuration');
+
+ConsoleRunner::addCommands($cli);
+
 $commands = $container->get('config')['console']['commands'];
 
 foreach ($commands as $command) {
     $cli->add($container->get($command));
 }
+
 $cli->run();