+++ src/App/Console/Command/FixtureCommand.php
@@ -1,0 +1,50 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Console\Command;
+
+use Doctrine\Common\DataFixtures\Executor\ORMExecutor;
+use Doctrine\Common\DataFixtures\Loader;
+use Doctrine\Common\DataFixtures\Purger\ORMPurger;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\Console as Console;
+
+class FixtureCommand extends Console\Command\Command
+{
+    private EntityManagerInterface $em;
+    private string $path;
+
+    public function __construct(EntityManagerInterface $em, string $path)
+    {
+        parent::__construct();
+        $this->em = $em;
+        $this->path = $path;
+    }
+
+    protected function configure(): void
+    {
+        $this
+            ->setName('fixtures:load')
+            ->setDescription('Load fixtures');
+    }
+
+    protected function execute(Console\Input\InputInterface $input, Console\Output\OutputInterface $output): int
+    {
+        $output->writeln('<comment>Loading fixtures</comment>>');
+        $loader = new Loader();
+        $loader->loadFromDirectory($this->path);
+
+        $executor = new ORMExecutor($this->em, new ORMPurger());
+
+        $executor->setLogger(function ($message) use ($output) {
+            $output->writeln($message);
+        });
+
+        $executor->execute($loader->getFixtures());
+
+        $output->writeln('<info>Done!</info>');
+
+        return 0;
+    }
+}